backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
tensorflow==2.7.0
numpy==1.19.5
Pillow==8.4.0
langchain-google-genai==0.0.11
google-generativeai==0.3.2
python-dotenv==1.0.0
